Supporting unsigned metrics in FASER & converting file writer metrics back to unsigned

See merge request faser/online/faser-daq!159

+++ src/Modules/FileWriterFaser/FileWriterFaserModule.hpp
@@ -75,11 +75,11 @@ private:
   using Context = std::tuple<PayloadQueue, ThreadContext>;
 
   struct Metrics {
-    std::atomic<int> bytes_written = 0;
-    std::atomic<int> events_received = 0;
-    std::atomic<int> files_written = 0;
-    std::atomic<int> payload_queue_size = 0;
-    std::atomic<int> payload_size = 0;
+    std::atomic<size_t> bytes_written = 0;
+    std::atomic<size_t> events_received = 0;
+    std::atomic<size_t> files_written = 0;
+    std::atomic<size_t> payload_queue_size = 0;
+    std::atomic<size_t> payload_size = 0;
   };
 
   size_t m_buffer_size;
